backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-oauth-toolkit
django-filter
pandas
python-dotenv
razorpay
twilio
firebase-admin
Pillow
python-magic
gunicorn
whitenoise
